Fix thinko in px_crypt_shacrypt().
We must deal with complete hashed password inputs. Originally
we just assumed that we get a salt with options passed down
to us, but i didn't read the API specification for crypt()
close enough:

It also supports the _whole_ password hash passed to it, so
we need to be careful to really just extract the salt string and
nothing more. With this, the caller can do

SELECT 'password' = crypt('password', 'shacrypted pw hash');

which meets the intended crypt() call criteria.