Follow up fixes for commit 749a9e20c9790006f3af47f7a8faf4ad8dc358d9.
This removes the platform dependant use of isalpha() and friends and replaces
the check for supported letters in the salt string to just check against
plain ASCII letters. While this makes the use of salt letters more strict,
we expect this to make the code more robust.

This fixes a careless use-pointer-after-pfree spotted by Andres Freund, too.